+++ tests/testthat/test-adasyn_impl.R
@@ -49,6 +49,10 @@ test_that("bad args", {
   )
   expect_snapshot(
     error = TRUE,
+    adasyn(circle_example, var = "class")
+  )
+  expect_snapshot(
+    error = TRUE,
     adasyn(circle_example_num, var = "class", k = 0)
   )
   expect_snapshot(