+++ tests/testthat/test-smote_impl.R
@@ -95,6 +95,17 @@ test_that("bad args", {
   )
   expect_snapshot(
     error = TRUE,
+    smote(circle_example, var = "class")
+  )
+  expect_snapshot(
+    error = TRUE,
+    smote(
+      data.frame(x = 1:4, a = letters[1:4], b = letters[1:4], class = factor(1:4)),
+      var = "class"
+    )
+  )
+  expect_snapshot(
+    error = TRUE,
     smote(circle_example, var = "class", k = 0)
   )
   expect_snapshot(